Initial commit, adding new submission_data route and adding revalidation_threshold to metadata
